Removes deprecated DishCard component.
