+++ src/components/MenuSection.client.jsx
@@ -1,6 +1,4 @@
 import React from 'react';
-// import CardFilters from "./CardFilters.client";
-// import DishCard from "./DishCard";
 import iconEdit from "../assets/icon-edit.png";
 import iconPlus from "../assets/icon-plus-alt.png";
 import iconArrowRight from "../assets/arrow-right.png"
@@ -11,7 +9,6 @@ import CardFilters from './CardFilters.client';
 import DishCard from './DishCard.client';
 import Modal from 'react-modal/lib/components/Modal';
 import { prepModSubTitles } from '../lib/utils';
-import DishCardFunc from './DishCardFunc.client';
 
 export default class MenuSection extends React.Component {
 
@@ -156,7 +153,7 @@ export default class MenuSection extends React.Component {
 
                 return (
                     <div className="dish-card-item" key={choice.title}>
-                        <DishCardFunc 
+                        <DishCard 
                             choice={choice}
                             pullStatus={this.pullStatus}
                             freeQuantityLimit={freeQuantityLimit}
